+++ fixtures/media-library/CHANGELOG.md
@@ -1,5 +1,21 @@
 # media-library-fixture
 
+## 1.0.13
+
+### Patch Changes
+
+- The following workspace dependencies were updated
+  - dependencies
+    - @sanity/cli bumped to 7.15.1
+
+## 1.0.12
+
+### Patch Changes
+
+- The following workspace dependencies were updated
+  - dependencies
+    - @sanity/cli bumped to 7.15.0
+
 ## 1.0.11
 
 ### Patch Changes